requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
google-api-python-client==2.108.0
openai==1.3.7
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
tinydb>=4.8.0
python-multipart>=0.0.6